# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Place Management API",
        "description": "API para gerenciar lugares",
        "contact": {
            "email": "domingosbreganha9@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/places": {
            "get": {
                "tags": [
                    "Places"
                ],
                "summary": "Lista todos os lugares",
                "operationId": "60e32ee62b6da9d31f5d04e3243a27e6",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtra os lugares pelo nome",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de lugares",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de lugares"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Place Name"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "place-name"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "City Name"
                                                    },
                                                    "state": {
                                                        "type": "string",
                                                        "example": "State Name"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-08T15:32:33Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-08-08T22:38:54Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sem lugares disponíveis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sem lugares disponíveis"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Places"
                ],
                "summary": "Cria um novo lugar",
                "operationId": "165c8fd925cfdb40f624ca61262d5ceb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "slug",
                                    "city",
                                    "state"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Place Name"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "City Name"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "State Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lugar criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lugar criado com sucesso"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Place Name"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "place-name"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "City Name"
                                                },
                                                "state": {
                                                    "type": "string",
                                                    "example": "State Name"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-08-08T12:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-08-08T12:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao criar lugar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao criar lugar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}